fix(clippy): three lint errors on main (manual_range_patterns + unnecessary_cast)
Fixes the red CI on main: 3 clippy errors landed with the UC8151D panel
and `spi_class_transfer` thunk from labwired-core#98 / #99. All
mechanical:

  - uc8151d_tricolor_290.rs:171 `0x21 | 0x22 | 0x23 | 0x24` →
    `0x21..=0x24` (manual_range_patterns)
  - uc8151d_tricolor_290.rs:181 `0x80 | 0x81 | 0x82` →
    `0x80..=0x82` (manual_range_patterns)
  - esp32s3/rom_thunks.rs:415 `(... & 0xFF) as u32` → `... & 0xFF`
    (unnecessary_cast; the LHS is already u32 after the &)

`cargo clippy --all-targets -- -D warnings` clean.
351 unit tests still pass.

+++ crates/core/src/peripherals/esp32s3/rom_thunks.rs
@@ -412,7 +412,7 @@ pub fn spi_class_transfer(cpu: &mut XtensaLx7, bus: &mut dyn Bus) -> SimResult<(
     let callinc = cpu.ps.callinc();
     let n = callinc * 4;
     let this = cpu.regs.read_logical(n + 2);
-    let byte = (cpu.regs.read_logical(n + 3) & 0xFF) as u32;
+    let byte = cpu.regs.read_logical(n + 3) & 0xFF;
 
     // Resolve SPI peripheral base. Prefer firmware's `_spi->dev` when
     // populated; fall back to SPI3 (the bus our SSD1680 model is
